+++ modules/logwatch.nix
@@ -10,23 +10,27 @@ let
   cfg = config.services.logwatch;
   types = lib.types;
 
-  packageConfig = {
+  extraConfig = {
     inherit (cfg)
-      mailer
       archives
+      mailer
       mailto
       mailfrom
       range
       detail
       services
-      customServices
-      extraFixup
-      extraPerl5Lib
-      extraPath
       ;
   };
 
-  logwatch = pkgs.callPackage ../packages/logwatch.nix { inherit packageConfig; };
+  defaultPackage = pkgs.callPackage ../packages/logwatch/package.nix {
+    logwatch-unwrapped = pkgs.callPackage ../packages/logwatch-unwrapped/package.nix { };
+    inherit extraConfig;
+    inherit (cfg)
+      customServices
+      extraPath
+      extraPerl5Lib
+      ;
+  };
 in
 {
   imports = [
@@ -34,10 +38,23 @@ in
       [ "services" "logwatch" "journalCtlEntries" ]
       [ "services" "logwatch" "customServices" ]
     )
+    (lib.mkRemovedOptionModule [
+      "services"
+      "logwatch"
+      "extraFixup"
+    ] "Use extraFixup or extraConfig for services, or override the unwrapped logwatch package instead")
   ];
 
   options.services.logwatch = {
     enable = lib.mkEnableOption "logwatch";
+
+    package = lib.mkOption {
+      type = lib.types.package;
+      default = defaultPackage;
+      description = ''
+        Which package to use for logwatch.
+      '';
+    };
 
     startAt = lib.mkOption {
       default = "*-*-* 4:00:00";
@@ -123,11 +140,6 @@ in
       type = types.listOf types.attrs;
       description = "What to watch";
     };
-    extraFixup = lib.mkOption {
-      default = "";
-      type = types.str;
-      description = "Arbitrary customization commands, added to the end of the fixupPhase";
-    };
     extraPath = lib.mkOption {
       default = [ ];
       type = types.listOf types.package;
@@ -141,12 +153,12 @@ in
   };
 
   config = lib.mkIf cfg.enable {
-    environment.systemPackages = [ logwatch ];
+    environment.systemPackages = [ cfg.package ];
     systemd.services.logwatch = {
       description = "Digests the system logs";
       serviceConfig = {
         Type = "oneshot";
-        ExecStart = ''${lib.getExe logwatch} --output mail'';
+        ExecStart = "${lib.getExe cfg.package} --output mail";
         PrivateTmp = true;
       };
     };
@@ -156,10 +168,11 @@ in
       wantedBy = [ "timers.target" ];
       after = [ "network.target" ];
       timerConfig = {
-        OnCalendar = if builtins.isString cfg.startAt then [ cfg.startAt ] else cfg.startAt;
+        OnCalendar = if lib.isString cfg.startAt then [ cfg.startAt ] else cfg.startAt;
         Persistent = cfg.persistent;
         RandomizedDelaySec = cfg.randomizedDelaySec;
       };
     };
   };
+  meta.maintainers = with lib.maintainers; [ sfrijters ];
 }